# docs/swagger.yaml
openapi: 3.0.0
info:
  title: TODO API
  description: A comprehensive REST API for managing tasks, workflows, and task management
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: http://prod-api.example.com/api/v1
    description: Production server
...
paths:
...
  /task-type:
    post:
      tags:
        - Task Type
      summary: Create a new task type
      description: Create a new task type
      operationId: createTaskType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskTypeRequest'
      responses:
        '201':
          description: Task type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskType'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Task Type
      summary: Get all task types
      description: Retrieve a list of all task types
      operationId: getAllTaskTypes
      responses:
        '200':
          description: List of task types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskType'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /task-type/{id}:
    get:
      tags:
        - Task Type
      summary: Get a specific task type
      description: Retrieve a single task type by its ID
      operationId: getTaskType
      parameters:
        - name: id
          in: path
          required: true
          description: Task Type ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskType'
        '404':
          description: Task type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Task Type
      summary: Update a task type
      description: Update an existing task type
      operationId: updateTaskType
      parameters:
        - name: id
          in: path
          required: true
          description: Task Type ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskTypeRequest'
      responses:
        '200':
          description: Task type updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskType'
        '404':
          description: Task type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Task Type
      summary: Delete a task type
      description: Delete a task type by its ID
      operationId: deleteTaskType
      parameters:
        - name: id
          in: path
          required: true
          description: Task Type ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task type deleted successfully
        '404':
          description: Task type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows:
    post:
      tags:
        - Workflows
      summary: Create a new workflow
      description: Create a new workflow with statuses
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Workflows
      summary: Get all workflows
      description: Retrieve a list of all workflows
      operationId: getAllWorkflows
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workflows/{id}:
    get:
      tags:
        - Workflows
      summary: Get a specific workflow
      description: Retrieve a single workflow by its ID
      operationId: getWorkflow
      parameters:
        - name: id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Workflows
      summary: Update a workflow
      description: Update an existing workflow
      operationId: updateWorkflow
      parameters:
        - name: id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Workflows
      summary: Delete a workflow
      description: Delete a workflow by its ID
      operationId: deleteWorkflow
      parameters:
        - name: id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Workflow deleted successfully
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    TaskStatus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        active:
          type: boolean
    CreateTaskStatusRequest:
      type: object
      required:
        - label
        - active
      properties:
        label:
          type: string
          example: "In Progress"
        active:
          type: boolean
          example: true
    UpdateTaskStatusRequest:
      type: object
      properties:
        label:
          type: string
        active:
          type: boolean
    TaskType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CreateTaskTypeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Bug"
    UpdateTaskTypeRequest:
      type: object
      properties:
        name:
          type: string
    Workflow:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        statuses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TaskStatus'
        author:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - statuses
      properties:
        name:
          type: string
          example: "Standard Workflow"
        statuses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CreateTaskStatusRequest'
    UpdateWorkflowRequest:
      type: object
      properties:
        name:
          type: string
        statuses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/UpdateTaskStatusRequest'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        email:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_DATABASE_CONNECTION"
        message:
          type: string
          example: "Failed to connect to database"